refactor: moving existing functions to datetime
It now makes sense for some existing functions to be in this new module
