+++ packages/core/src/awsService/iot/explorer/iotCertificateNode.ts
@@ -20,9 +20,10 @@ import { IotCertsFolderNode } from './iotCertFolderNode'
 import { IotThingNode } from './iotThingNode'
 import { IotPolicyCertNode } from './iotPolicyNode'
 import { getIcon } from '../../../shared/icons'
-import { formatLocalized, truncate } from '../../../shared/utilities/textUtilities'
+import { truncate } from '../../../shared/utilities/textUtilities'
 import { Settings } from '../../../shared/settings'
 import { ClassToInterfaceType } from '../../../shared/utilities/tsUtils'
+import { formatLocalized } from '../../../shared/datetime'
 
 const contextBase = 'awsIotCertificateNode'
 /**
